requirements.txt
aiodns==2.0.0
aiosmtpd==1.2.2
asgiref==3.2.10
atpublic==2.1.1
cffi==1.14.4
Django==3.1.2
django-ckeditor==6.0.0
django-countries==7.0
django-js-asset==1.2.2
dnspython==2.0.0
filelock==3.0.12
idna==2.10
Pillow==8.0.1
pycares==3.1.1
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
validate-email==1.3
verify-email==2.4.1
